Implementation/AIRAGSecurityModel/requirements.txt
# Expert-level Python dependencies for AIRAGSecurityModel
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.2
sentence-transformers==2.6.1
faiss-cpu==1.7.4
transformers==4.40.2
scikit-learn==1.4.2
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0